chore: split to cli and dev-server
