+++ packages/rspack-cli/bin/rspack
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require("../lib/index.js");
