+++ packages/rspack-cli/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@rspack/cli",
+  "version": "0.0.4",
+  "bin": {
+    "rspack": "./bin/rspack"
+  },
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
+  "scripts": {
+    "build": "rm -rf lib/ && tsc",
+    "dev": "tsc -w",
+    "test": ""
+  },
+  "files": [
+    "lib"
+  ],
+  "dependencies": {
+    "@rspack/core": "workspace:*",
+    "@rspack/dev-server": "workspace:*",
+    "commander": "9.4.0"
+  }
+}
