+++ packages/rspack-cli/tsconfig.json
@@ -5,7 +5,8 @@
     "moduleResolution": "node",
     "esModuleInterop": true,
     "outDir": "lib",
-    "declaration": true
+    "declaration": true,
+    "isolatedModules": true
   },
   "include": ["src"]
 }
