+++ packages/rspack-dev-server/package.json
@@ -2,23 +2,25 @@
   "name": "@rspack/dev-server",
   "description": "",
   "version": "0.0.4",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
   "scripts": {
-    "build": "rm -rf dist/ && tsc -p tsconfig.json",
-    "dev": "tsc -p tsconfig.json -w",
+    "build": "rm -rf lib/ && tsc",
+    "dev": "tsc -w",
     "test": "echo success"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/node": "^18.6.3",
     "tsm": "^2.2.2",
-    "typescript": "^4.7.4"
+    "@rspack/core": "workspace:*",
+    "typescript": "^4.7.4",
+    "@types/node": "16.11.7"
   },
   "dependencies": {
-    "@types/express": "^4.17.13",
-    "chalk": "^5.0.1",
     "express": "^4.18.1"
+  },
+  "peerDependencies": {
+    "@rspack/core": "*"
   }
 }
