+++ packages/rspack/package.json
@@ -10,7 +10,7 @@
     "test": "jest --verbose"
   },
   "files": [
-    "lib",
+    "lib"
   ],
   "devDependencies": {
     "@types/node": "^18.6.3",
@@ -30,11 +30,7 @@
   },
   "dependencies": {
     "@rspack/binding": "workspace:*",
-    "ws": "8.8.1",
     "chokidar": "3.5.3",
-    "open": "8.4.0",
-    "clipanion": "3.2.0-rc.11",
-    "sirv": "2.0.2",
     "@rspack/plugin-postcss": "workspace:^",
     "tapable": "2.2.1",
     "webpack-sources": "3.2.3",
